Table creation validation for install routine (https://github.com/woocommerce/woocommerce-blocks/pull/2287)
* Bump package version for testing

* New install routine and db version

* Revert "Bump package version for testing"

This reverts commit c3fee55a34e75429b731cd7a2cc07e841010df6d.

* Update bootstrap to ensure blocks is installed properly before tests are ran

* Restore errors after install

* Use enabled method instead of comparing db version

* Moved installer code to admin_init

* Tweak test bootstrap

+++ plugins/woocommerce-blocks/src/StoreApi/Utilities/ReserveStock.php
@@ -14,6 +14,29 @@ defined( 'ABSPATH' ) || exit;
  */
 final class ReserveStock {
 	/**
+	 * Is stock reservation enabled?
+	 *
+	 * @var boolean
+	 */
+	private $enabled = true;
+
+	/**
+	 * Constructor
+	 */
+	public function __construct() {
+		$this->enabled = get_option( 'wc_blocks_db_schema_version', 0 ) >= 260;
+	}
+
+	/**
+	 * Is stock reservation enabled?
+	 *
+	 * @return boolean
+	 */
+	protected function is_enabled() {
+		return $this->enabled;
+	}
+
+	/**
 	 * Query for any existing holds on stock for this item.
 	 *
 	 * @param \WC_Product $product Product to get reserved stock for.
@@ -23,6 +46,10 @@ final class ReserveStock {
 	 */
 	public function get_reserved_stock( \WC_Product $product, $exclude_order_id = 0 ) {
 		global $wpdb;
+
+		if ( ! $this->is_enabled() ) {
+			return 0;
+		}
 
 		// phpcs:ignore WordPress.DB.PreparedSQL.InterpolatedNotPrepared, WordPress.DB.PreparedSQL.NotPrepared
 		return (int) $wpdb->get_var( $this->get_query_for_reserved_stock( $product->get_stock_managed_by_id(), $exclude_order_id ) );
@@ -39,7 +66,7 @@ final class ReserveStock {
 	public function reserve_stock_for_order( \WC_Order $order, $minutes = 0 ) {
 		$minutes = $minutes ? $minutes : (int) get_option( 'woocommerce_hold_stock_minutes', 60 );
 
-		if ( ! $minutes ) {
+		if ( ! $minutes || ! $this->is_enabled() ) {
 			return;
 		}
 
@@ -94,6 +121,10 @@ final class ReserveStock {
 	 */
 	public function release_stock_for_order( \WC_Order $order ) {
 		global $wpdb;
+
+		if ( ! $this->is_enabled() ) {
+			return;
+		}
 
 		$wpdb->delete(
 			$wpdb->wc_reserved_stock,
